requirements.txt
gradio>=5.0.0
openai>=1.0.0
supabase
python-dotenv
